+++ CHANGELOG.md
@@ -12,7 +12,14 @@ You can also check the
 # Unreleased
 
 - Features
+
+  - Added stacked horizontal charts for power stability to the sunshine charts
+
+- Storybook
+
+  - Added stacked horizontal chart to storybook
   - SAIDI / SAIFI indicators can be shown on the map (behind "sunshine" flag)
+  - 
 - Fixes
 
   - Changes to Docker image to be able to use duckdb (alpine based to debian based)